requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pywebview==5.0
polars==0.20.0
xlsxwriter==3.1.9
fastexcel==0.10.0
pymupdf==1.23.21
python-docx==1.1.0
onnxruntime==1.17.0
rapidocr_onnxruntime>=1.3.0
opencv-python-headless==4.9.0.80
Pillow==10.2.0
playwright==1.41.0
playwright-stealth==1.0.6
watchdog==4.0.0
psutil==5.9.8
rapidfuzz==3.6.1
py-cpuinfo==9.0.0
cryptography==42.0.2
